requirements.txt
face_recognition==1.3.0
Flask==2.0.2
Flask_Cors==3.0.10
numpy==1.21.4
opencv-python-headless==4.5.4.60
Pillow==8.4.0
gunicorn
